Palindrome Integer
An alternative way to check if an integer is a palindrome. This method stores the last digit of the integer into a temp variable. Then each iteration the last digit of the temp is stored into the remainder variable. Then the remainder number is added to the reverse variable to eventually compare if the integer is a palindrome.